windvector to single precision

+++ himan-plugins/include/windvector.cuh
@@ -22,7 +22,7 @@ enum HPWindVectorTargetType
 
 namespace windvector_cuda
 {
-void RunCuda(const std::shared_ptr<const plugin_configuration> conf, std::shared_ptr<info<double>> myTargetInfo,
+void RunCuda(const std::shared_ptr<const plugin_configuration> conf, std::shared_ptr<info<float>> myTargetInfo,
              const param& UParam, const param& VParam, HPWindVectorTargetType itsTargetType);
 
 }  // namespace windvector_cuda
